init layerscale gamma in bf16

+++ train_gpt.py
@@ -216,7 +216,7 @@ class LayerScale(nn.Module):
     ) -> None:
         super().__init__()
         self.init_values = init_values
-        self.gamma = nn.Parameter(torch.empty(dim))
+        self.gamma = nn.Parameter(torch.empty(dim), dtype=torch.bfloat16)
         self.reset_parameters()
 
     def reset_parameters(self):
